src/ia_utils/core/parser.py
"""hOCR and metadata parsing utilities."""

from typing import List, Dict, Any, Optional, Tuple
import xml.etree.ElementTree as ET
from bs4 import BeautifulSoup, Tag


def parse_metadata(meta_bytes: bytes) -> dict:
    """Parse meta.xml bytes."""
    # TODO: Implement metadata parsing
    pass


def parse_files(files_bytes: bytes) -> list:
    """Parse files.xml bytes."""
    # TODO: Implement files parsing
    pass


def parse_hocr(hocr_bytes: bytes) -> list:
    """Parse hOCR HTML bytes and extract text blocks."""
    # TODO: Implement hOCR parsing
    pass
